% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr5000.R
\docType{data}
\name{dplyr5000}
\alias{dplyr5000}
\title{Data from the 5000 first travis builds of dplyr}
\format{tibble with columns I'll document later}
\usage{
dplyr5000
}
\description{
Data from the 5000 first travis builds of dplyr
}
\keyword{datasets}
